hw3/requirements.txt
functions-framework==3.2.1
flask==2.2.5
google-cloud-error-reporting==1.9.1
google-cloud-storage==2.11.0